Fixes bug that prevents unmarshaling valid LLRP messages when they have an "empty" TLV.
Fixes DE925.

The original unmarshaling code unmarshals TLV parameters
so long as the payload has more than 4 bytes,
which is the size of a TLV parameter header.
However, this check should be for ">=", not ">",
as in some cases, an empty TLV header is a valid TLV.
This fixes the generator code
and adds a test case with the data that lead to the defect.